Fix: cart update & checkout
Cart view was only rendering on load. I added a cartView.render call
in the cart change event. Also checkout completion was failing because
removeEventListener was being called without reference to the listener
function. I converted the anonymous listener function into a named
function and passed reference to it in removeEventListener.